Fixed bug in assert statements
Was comparing two different types, added size_t  declarations to fix this.

+++ packages/containers/src/ops/tpetra/containers_mvec_prod_vec.hpp
@@ -82,7 +82,7 @@ void product(const mvec_type & mvA, const vec_type & vecB, res_type & C){
   // how many vectors are in mvA
   const auto numVecs = mvA.globalNumVectors();
   // size of vecB
-  assert(numVecs == vecB.size());
+  assert(size_t(numVecs) == size_t(vecB.size()));
 
   // my number of rows
   const auto myNrows = mvA.localLength();
@@ -171,7 +171,7 @@ void product(const mvec_type & mvA, const vec_type & vecB, res_type & C){
   // how many vectors are in mvA
   const auto numVecs = mvA.globalNumVectors();
   // size of vecB
-  assert(size_t(numVecs) == vecB.size());
+  assert(size_t(numVecs) == size_t(vecB.size()));
 
   // my number of rows
   const auto myNrows = mvA.localLength();
